backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
jwt==1.4.0
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
peewee==3.19.0
pymysql==1.1.0
playwright==1.57.0
pluggy==1.6.0
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlite-fts4==1.0.3
sqlite-utils==3.39
starlette==0.50.0
tabulate==0.9.0
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.5
wsproto==1.3.2
